Fix test bug on printHTML()
