+++ src/modules/printHTML.js
@@ -6,6 +6,7 @@ import { switchTask } from './switchTask.js';
 
 const printHTML = (description, status, index) => {
   const taskShelf = document.querySelector('#to-do-list__shelf');
+ 
   taskShelf.insertAdjacentHTML(
     'beforeend',
     `
